requirements.txt
# This file is used by Poetry to specify dependencies of your project
# See https://python-poetry.org/docs/dependency-specification/

qrcode
Pillow
pyzbar
opencv-python
